requirements.txt
streamlit==1.21.0
Pillow==9.5.0
transformers==4.28.1
torch>=1.9.0
